Switch to a layout on debugger start.
Add debugger setting to select a layout that is switched to on
debugger start. Existing layout is also stored in a temporary file and
is restored on debugger stop. The temporary file is then deleted. If
the temporary layout file exists on FD startup, it is assumed that
FD exited while debugger is still running, and restored to old layout
on startup.

Add LayoutSelectorEditor UI element for selecting an existing layout.
Because of this, FlashDebugger plugin now depends on LayoutManager plugin.