More AIR Wizard improvements
Added missing properties SoftKeyboardBehavior and DepthAndStencil.
Added support to specify minimum iOS and Android target OS version.
Added detection of ANEs defined in the project, and if removed from the
wizard, the asset is also removed from the external libraries list.
Disabled edition of extension ids if the ANE has been added to the
project.