Rename and Move commands improvements
Fixed tons of errors and unsupported cases, mainly in the Move command.
Rename command left modified files unsaved, while Move command saved
them, the latter is the common approach taken, so made Rename to behave
the same.
The Move command tried to keep modified files open, but the code was
wrong, I fixed the initial check but left it commented, I think it
pollutes the user interfaces when the refactoring affects many files. It
would require some further changes.
When the Move and Rename commands finish the affected files that were
initially open are kept open.