Some more work
Fixed MethodCallTip display if it's already visible.
Some more decoupling from Scintilla.
Changes to the Immediate Panel to use CompletionListControl and serve as
an example, as well, and help define the needs we may have.
Changed behaviour of navigation through immediate panel, and allow to
better navigate through the previous entered commands with the keyboard.