Fixed an inconsistency in the settings filter
The description of a property was only being taken into account when filtering the currently selected plugin's properties, but not when repopulating the plugin list itself.
This led to effects like the CodeFormatter plugin not appearing for the "Trim " filter String, even though it's contained in the Trim trailing spaces property's description.