Update HaxeProjects when the classpath is rebuilt
Previously, FD would not recognize any version change of haxelibs (e.g. when doing a haxelib set on the command line or installing a new version), even when manually refreshing the project - the project had to be reloaded entirely.
There might be a better solution that automatically detects changes, like a FileSystemWatcher for the .current files.