Escaped quote does not override closing quote
By checking that the style at current position equals to `CPP.STRINGEOL`, you can ensure that the currently entered quote terminates the string literal - an escaped quote will still have the style `CPP.STRING` or `CPP.CHARACTER`. 
However it does not work at the start of file or at the end of file.