Scale the max width of the CompletionList
Resolves #1138

There was a max width defined for the completion list of 400 pixels regardless of scaling. Changed to use ScaleHelper.Scale so that this max width will be larger on HiDPI displays. Since there is a max width, there is still the chance that long namespaces will cause suggestions to be cut off, however the completion list should render consistently despite different DPIs.