Prevent dynamic persist panel from force-activating
Panels created using CreateDockablePanel() are expected to be created
during plugin initialization, which is right before
LayoutManager.BuildLayoutSystems(). That method ensures that all panels
are in the correct place, and shows unregistered panels as floating,
essentially making the call to Show() in CreateDockablePanel()
unnecessary.
Panels created using CreateDynamicPersistDockablePanel() are also called
LayoutManager.SetContentLayout(), which ensures the panel is in the
correct place if registered. Adding Show() at the end ensures the panel
is shown when the panel information isn't stored in the layout file.