Follow-up fix for invisible progress bars (#2896)
* Use ProgressBarEx as the progress bar type

If we don't declare progressBar as a ProgressBarEx, the bar does not show up in the UI.

* Use ProgressBarEx as the progress bar type

If we don't declare progressBar as a ProgressBarEx, the bar does not show up in the UI.

* Follow-up fix for invisible progress bars

The previous fix did not work in all cases, leaving the progress bars not rendering. This should fix it properly, as using the ControlStyles.UserPaint style was unreliable, but catching the WM_PAINT message works well.